# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SISTEMA EDUPOLI COMPONENTE BACKEND",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://proyectoedupoli.herokuapp.com/"
        }
    ],
    "paths": {
...
        "/api/v1/comentarios/sistema/cambio/{semestres}": {
            "post": {
                "tags": [
                    "Gestion Comentarios"
                ],
                "summary": "Crear comentario",
                "description": "Ingrese la informacion para crear un comentario",
                "operationId": "Crear comentarios en relacion al semestres",
                "parameters": [
                    {
                        "name": "semestres",
                        "in": "path",
                        "description": "Id del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "comentario"
                                ],
                                "properties": {
                                    "comentario": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/comentarios/sistema": {
            "get": {
                "tags": [
                    "Gestion Comentarios"
                ],
                "summary": "Lista de comentarios",
                "description": "Lista de comentarios",
                "operationId": "Comentarios",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/comentarios/sistema/{id}": {
            "get": {
                "tags": [
                    "Gestion Comentarios"
                ],
                "summary": " Información de un comentario según su id",
                "description": "Retorna la información de un comentario",
                "operationId": "Ver comentario por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del semestre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gestion Comentarios"
                ],
                "summary": "Eliminar Comentarios",
                "operationId": "9c28413474daa89fe0f225dcebfc0953",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/comentarios/sistema/actualizar/{id}": {
            "post": {
                "tags": [
                    "Gestion Comentarios"
                ],
                "summary": "Actualizar comentario",
                "description": " Ingrese el comentario a actualizarce",
                "operationId": " Actualizar Comentarios",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "comentario": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backendedupoli.test/oauth/authorize",
                        "tokenUrl": "http://backendedupoli.test/oauth/token",
                        "refreshUrl": "http://backendedupoli.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}